refactor: PostContext, UserContext 추가 및 모달 컴포넌트 분리
