--- src/features/post/ui/AddPostModal.tsx
@@ -1,39 +1,0 @@
-import { DialogContent, DialogTitle } from "@radix-ui/react-dialog"
-import { Button, DialogHeader, Input, Textarea } from "../../../shared/ui"
-import { NewPost } from "../../../entities/post/model/types"
-
-interface Props {
-  newPost: NewPost
-  setNewPost: (newPost: NewPost) => void
-  addPost: () => void
-}
-
-export const AddPostModal = ({ newPost, setNewPost, addPost }: Props) => {
-  return (
-    <DialogContent>
-      <DialogHeader>
-        <DialogTitle>새 게시물 추가</DialogTitle>
-      </DialogHeader>
-      <div className="space-y-4">
-        <Input
-          placeholder="제목"
-          value={newPost.title}
-          onChange={(e) => setNewPost({ ...newPost, title: e.target.value })}
-        />
-        <Textarea
-          rows={30}
-          placeholder="내용"
-          value={newPost.body}
-          onChange={(e) => setNewPost({ ...newPost, body: e.target.value })}
-        />
-        <Input
-          type="number"
-          placeholder="사용자 ID"
-          value={newPost.userId}
-          onChange={(e) => setNewPost({ ...newPost, userId: Number(e.target.value) })}
-        />
-        <Button onClick={addPost}>게시물 추가</Button>
-      </div>
-    </DialogContent>
-  )
-}
