+++ src/features/post/ui/PostDetailModal.tsx
@@ -1,24 +1,26 @@
-import { DialogContent, DialogTitle } from "@radix-ui/react-dialog"
-import { DialogHeader } from "../../../shared/ui"
+import { Dialog, DialogContent, DialogHeader, DialogTitle } from "../../../shared/ui"
 import { highlightText } from "../../../shared/lib/highlightText"
-import { Post } from "../../../entities/post/model/types"
 import { Comments } from "../../comment/ui/Comments"
+import { usePostsContext } from "../../../shared/model/PostContext"
 
 interface Props {
-  selectedPost: Post
   searchQuery: string
 }
 
-export const PostDetailModal = ({ selectedPost, searchQuery }: Props) => {
+export const PostDetailModal = ({ searchQuery }: Props) => {
+  const { showPostDetailDialog, setShowPostDetailDialog, selectedPost } = usePostsContext()
+
   return (
-    <DialogContent className="max-w-3xl">
-      <DialogHeader>
-        <DialogTitle>{highlightText(selectedPost?.title, searchQuery)}</DialogTitle>
-      </DialogHeader>
-      <div className="space-y-4">
-        <p>{highlightText(selectedPost?.body, searchQuery)}</p>
-        {selectedPost && <Comments postId={selectedPost.id} searchQuery={searchQuery} />}
-      </div>
-    </DialogContent>
+    <Dialog open={showPostDetailDialog} onOpenChange={setShowPostDetailDialog}>
+      <DialogContent className="max-w-3xl">
+        <DialogHeader>
+          <DialogTitle>{highlightText(selectedPost?.title, searchQuery)}</DialogTitle>
+        </DialogHeader>
+        <div className="space-y-4">
+          <p>{highlightText(selectedPost?.body, searchQuery)}</p>
+          {selectedPost && <Comments postId={selectedPost.id} searchQuery={searchQuery} />}
+        </div>
+      </DialogContent>
+    </Dialog>
   )
 }
